+++ buttplug/src/server/device/protocol/svakom_v4.rs
@@ -53,7 +53,7 @@ impl ProtocolHandler for SvakomV4 {
         actuator,
         0x00,
         if scalar == 0 { 0x00 } else { 0x01 },
-        scalar as u8,
+        scalar,
       ]
       .to_vec(),
       false,
